% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildFilesJSON.R
\name{BuildFilesJSON}
\alias{BuildFilesJSON}
\title{BuildFilesJSON}
\usage{
BuildFilesJSON(conf.setup)
}
\arguments{
\item{conf.setup}{field a list of Lattes CV XML files}
}
\description{
BuildFilesJSON
}
