% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_json.R
\name{BuildJson}
\alias{BuildJson}
\title{BuildJson}
\usage{
BuildJson(pessoa, publicacoes, orientacoes, senioridade)
}
\arguments{
\item{pessoa}{field a list of Lattes CV XML files.}

\item{publicacoes}{field a list of Lattes CV XML files.}

\item{orientacoes}{field a list of Lattes CV XML files.}

\item{senioridade}{field a list of Lattes CV XML files.}
}
\description{
BuildJson
}
