% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCAR_AREAS_DE_ATUACAO.R
\name{BUSCAR_AREAS_DE_ATUACAO}
\alias{BUSCAR_AREAS_DE_ATUACAO}
\title{BUSCAR_AREAS_DE_ATUACAO}
\usage{
BUSCAR_AREAS_DE_ATUACAO(doc)
}
\arguments{
\item{doc}{a document}
}
\description{
BUSCAR_AREAS_DE_ATUACAO
}
