% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCAR_ARTIGOS_ACEITOS_PARA_PUBLICACAO.R
\name{BUSCAR_ARTIGOS_ACEITOS_PARA_PUBLICACAO}
\alias{BUSCAR_ARTIGOS_ACEITOS_PARA_PUBLICACAO}
\title{BUSCAR_ARTIGOS_ACEITOS_PARA_PUBLICACAO}
\usage{
BUSCAR_ARTIGOS_ACEITOS_PARA_PUBLICACAO(doc)
}
\arguments{
\item{doc}{a list of Lattes CV XML files}
}
\description{
BUSCAR_ARTIGOS_ACEITOS_PARA_PUBLICACAO
}
