% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCAR_ARTIGOS_PUBLICADOS.R
\name{BUSCAR_ARTIGOS_PUBLICADOS}
\alias{BUSCAR_ARTIGOS_PUBLICADOS}
\title{BUSCAR_ARTIGOS_PUBLICADOS}
\usage{
BUSCAR_ARTIGOS_PUBLICADOS(doc)
}
\arguments{
\item{doc}{a list of Lattes CV XML files}
}
\description{
BUSCAR_ARTIGOS_PUBLICADOS
}
