% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCAR_CAPITULOS_DE_LIVROS_PUBLICADOS.R
\name{BUSCAR_CAPITULOS_DE_LIVROS_PUBLICADOS}
\alias{BUSCAR_CAPITULOS_DE_LIVROS_PUBLICADOS}
\title{BUSCAR_CAPITULOS_DE_LIVROS_PUBLICADOS}
\usage{
BUSCAR_CAPITULOS_DE_LIVROS_PUBLICADOS(doc)
}
\arguments{
\item{doc}{a list of Lattes CV XML files}
}
\description{
BUSCAR_CAPITULOS_DE_LIVROS_PUBLICADOS
}
