% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCAR_DEMAIS_TIPOS_DE_PRODUCAO_BIBLIOGRAFICA.R
\name{BUSCAR_DEMAIS_TIPOS_DE_PRODUCAO_BIBLIOGRAFICA}
\alias{BUSCAR_DEMAIS_TIPOS_DE_PRODUCAO_BIBLIOGRAFICA}
\title{BUSCAR_DEMAIS_TIPOS_DE_PRODUCAO_BIBLIOGRAFICA}
\usage{
BUSCAR_DEMAIS_TIPOS_DE_PRODUCAO_BIBLIOGRAFICA(doc)
}
\arguments{
\item{doc}{a list of Lattes CV XML files}
}
\description{
BUSCAR_DEMAIS_TIPOS_DE_PRODUCAO_BIBLIOGRAFICA
}
