% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCAR_LIVROS_PUBLICADOS_OU_ORGANIZADOS.R
\name{BUSCAR_LIVROS_PUBLICADOS_OU_ORGANIZADOS}
\alias{BUSCAR_LIVROS_PUBLICADOS_OU_ORGANIZADOS}
\title{BUSCAR_LIVROS_PUBLICADOS_OU_ORGANIZADOS}
\usage{
BUSCAR_LIVROS_PUBLICADOS_OU_ORGANIZADOS(doc)
}
\arguments{
\item{doc}{a list of Lattes CV XML files}
}
\description{
BUSCAR_LIVROS_PUBLICADOS_OU_ORGANIZADOS
}
