% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/BUSCAR_ORIENTACOES_CONCLUIDAS_PARA_POS_DOUTORADO.R
\name{BUSCAR_ORIENTACOES_CONCLUIDAS_PARA_POS_DOUTORADO}
\alias{BUSCAR_ORIENTACOES_CONCLUIDAS_PARA_POS_DOUTORADO}
\title{BUSCAR_ORIENTACOES_CONCLUIDAS_PARA_POS_DOUTORADO}
\usage{
BUSCAR_ORIENTACOES_CONCLUIDAS_PARA_POS_DOUTORADO(doc)
}
\arguments{
\item{files}{a list of Lattes CV XML files}
}
\description{
BUSCAR_ORIENTACOES_CONCLUIDAS_PARA_POS_DOUTORADO
}
