% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/BUSCAR_ORIENTACOES_EM_ANDAMENTO_DE_INICIACAO_CIENTIFICA.R
\name{BUSCAR_ORIENTACOES_EM_ANDAMENTO_DE_INICIACAO_CIENTIFICA}
\alias{BUSCAR_ORIENTACOES_EM_ANDAMENTO_DE_INICIACAO_CIENTIFICA}
\title{BUSCAR_ORIENTACOES_EM_ANDAMENTO_DE_INICIACAO_CIENTIFICA}
\usage{
BUSCAR_ORIENTACOES_EM_ANDAMENTO_DE_INICIACAO_CIENTIFICA(doc)
}
\arguments{
\item{doc}{a list of Lattes CV XML files}
}
\description{
BUSCAR_ORIENTACOES_EM_ANDAMENTO_DE_INICIACAO_CIENTIFICA
}
