% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCAR_ORIENTACOES_EM_ANDAMENTO_DOUTORADO.R
\name{BUSCAR_ORIENTACOES_EM_ANDAMENTO_DOUTORADO}
\alias{BUSCAR_ORIENTACOES_EM_ANDAMENTO_DOUTORADO}
\title{BUSCAR_ORIENTACOES_EM_ANDAMENTO_DOUTORADO}
\usage{
BUSCAR_ORIENTACOES_EM_ANDAMENTO_DOUTORADO(doc)
}
\arguments{
\item{doc}{a list of Lattes CV XML files}
}
\description{
BUSCAR_ORIENTACOES_EM_ANDAMENTO_DOUTORADO
}
