% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCAR_ORIENTACOES_EM_ANDAMENTO_GRADUACAO.R
\name{BUSCAR_ORIENTACOES_EM_ANDAMENTO_GRADUACAO}
\alias{BUSCAR_ORIENTACOES_EM_ANDAMENTO_GRADUACAO}
\title{BUSCAR_ORIENTACOES_EM_ANDAMENTO_GRADUACAO}
\usage{
BUSCAR_ORIENTACOES_EM_ANDAMENTO_GRADUACAO(doc)
}
\arguments{
\item{doc}{a list of Lattes CV XML files}
}
\description{
BUSCAR_ORIENTACOES_EM_ANDAMENTO_GRADUACAO
}
