% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCAR_ORIENTACOES_EM_ANDAMENTO_MESTRADO.R
\name{BUSCAR_ORIENTACOES_EM_ANDAMENTO_MESTRADO}
\alias{BUSCAR_ORIENTACOES_EM_ANDAMENTO_MESTRADO}
\title{BUSCAR_ORIENTACOES_EM_ANDAMENTO_MESTRADO}
\usage{
BUSCAR_ORIENTACOES_EM_ANDAMENTO_MESTRADO(doc)
}
\arguments{
\item{doc}{a list of Lattes CV XML files}
}
\description{
BUSCAR_ORIENTACOES_EM_ANDAMENTO_MESTRADO
}
