% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCAR_OUTRAS_ORIENTACOES_CONCLUIDAS.R
\name{BUSCAR_OUTRAS_ORIENTACOES_CONCLUIDAS}
\alias{BUSCAR_OUTRAS_ORIENTACOES_CONCLUIDAS}
\title{BUSCAR_OUTRAS_ORIENTACOES_CONCLUIDAS}
\usage{
BUSCAR_OUTRAS_ORIENTACOES_CONCLUIDAS(doc)
}
\arguments{
\item{doc}{a list of Lattes CV XML files}
}
\description{
BUSCAR_OUTRAS_ORIENTACOES_CONCLUIDAS
}
