% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCAR_TEXTOS_EM_JORNAIS_OU_REVISTAS.R
\name{BUSCAR_TEXTOS_EM_JORNAIS_OU_REVISTAS}
\alias{BUSCAR_TEXTOS_EM_JORNAIS_OU_REVISTAS}
\title{BUSCAR_TEXTOS_EM_JORNAIS_OU_REVISTAS}
\usage{
BUSCAR_TEXTOS_EM_JORNAIS_OU_REVISTAS(doc)
}
\arguments{
\item{doc}{a list of Lattes CV XML files}
}
\description{
BUSCAR_TEXTOS_EM_JORNAIS_OU_REVISTAS
}
