% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BUSCAR_TRABALHOS_EM_EVENTOS.R
\name{BUSCAR_TRABALHOS_EM_EVENTOS}
\alias{BUSCAR_TRABALHOS_EM_EVENTOS}
\title{BUSCAR_TRABALHOS_EM_EVENTOS}
\usage{
BUSCAR_TRABALHOS_EM_EVENTOS(doc)
}
\arguments{
\item{doc}{a list of Lattes CV XML files}
}
\description{
BUSCAR_TRABALHOS_EM_EVENTOS
}
