% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construir_dado_grafo.R
\name{ConstruirDadoGrafo}
\alias{ConstruirDadoGrafo}
\title{ConstruirDadoGrafo}
\usage{
ConstruirDadoGrafo(dadosBasicos, publicacoesPorAno, configuracao, default)
}
\arguments{
\item{dadosBasicos}{a list of Lattes CV XML files.}

\item{publicacoesPorAno}{field a list of Lattes CV XML files.}

\item{configuracao}{field a list of Lattes CV XML files.}

\item{default}{field a list of Lattes CV XML files.}
}
\description{
ConstruirDadoGrafo
}
