% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_profile_to_json.R
\name{ConvertDataProfileToJson}
\alias{ConvertDataProfileToJson}
\title{ConvertDataProfileToJson}
\usage{
ConvertDataProfileToJson(dadoBasico, publicacoesPorAno, orientacoesPorAno)
}
\arguments{
\item{dadoBasico}{a list of Lattes CV XML files.}

\item{publicacoesPorAno}{field a list of Lattes CV XML files.}

\item{orientacoesPorAno}{field a list of Lattes CV XML files.}
}
\description{
ConvertDataProfileToJson
}
