% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_researcheres_by_area_to_json.R
\name{ConvertDataResearcheresByAreaToJson}
\alias{ConvertDataResearcheresByAreaToJson}
\title{ConvertDataResearcheresByAreaToJson}
\usage{
ConvertDataResearcheresByAreaToJson(ids, production.by.researcher)
}
\arguments{
\item{ids, }{production.by.researcher a list of Lattes CV XML files}
}
\description{
ConvertDataResearcheresByAreaToJson
}
