% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_to_json.R
\name{ConvertDataToJson}
\alias{ConvertDataToJson}
\title{ConvertDataToJson}
\usage{
ConvertDataToJson(conf.JSON, type.production, default,
  production.by.year.and.type)
}
\arguments{
\item{conf.JSON, }{type.production, default, production.by.year.and.type a list of Lattes CV XML files.}

\item{type.production}{field a list of Lattes CV XML files.}

\item{default}{field a list of Lattes CV XML files.}

\item{production.by.year.and.type}{field a list of Lattes CV XML files.}
}
\description{
ConvertDataToJson
}
