% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CREATE_GRAPH.R
\name{CreateGraph}
\alias{CreateGraph}
\title{CreateGraphC}
\usage{
CreateGraph(profiles, publications)
}
\arguments{
\item{profiles, publications}{json of profiles and publications}
}
\value{
return graphs
}
\description{
This function convert a set of Lattes CV XML files to a hash table indexed by #'id_LATTES each researcher
}
