% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{Execute}
\alias{Execute}
\title{Execute}
\usage{
Execute(ids, dado, conf, tipo, doTipo, levenshtein)
}
\arguments{
\item{ids, }{dado, conf, tipo, doTipo,levenshtein a list of Lattes CV XML files}
}
\description{
Execute
}
