% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expertise_area_to_json.R
\name{ExpertiseAreaToJson}
\alias{ExpertiseAreaToJson}
\title{ExpertiseAreaToJson}
\usage{
ExpertiseAreaToJson(pessoa)
}
\arguments{
\item{pessoa}{a list of Lattes CV XML files.}
}
\description{
ExpertiseAreaToJson
}
