% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{Filter}
\alias{Filter}
\title{Filter}
\usage{
Filter(ids, h, parametros, tipo)
}
\arguments{
\item{ids}{tipo a list of Lattes CV XML files.}

\item{h}{field a list of Lattes CV XML files.}

\item{parametros}{field a list of Lattes CV XML files.}

\item{tipo}{field a list of Lattes CV XML files.}
}
\description{
Filter
}
