% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_areas_name.R
\name{GetAreasName}
\alias{GetAreasName}
\title{GetAreasName}
\usage{
GetAreasName(ids.researcher, production.by.researcher)
}
\arguments{
\item{ids.researcher}{a list of Lattes CV XML files.}

\item{production.by.researcher}{field a list of Lattes CV XML files.}
}
\description{
GetAreasName
}
