% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_parameters.R
\name{GetDefaultParameters}
\alias{GetDefaultParameters}
\title{GetDefaultParameters}
\usage{
GetDefaultParameters()
}
\arguments{
\item{files}{a list of Lattes CV XML files}
}
\description{
GetDefaultParameters
}
