% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_researchers_by_area.R
\name{GetResearchersByArea}
\alias{GetResearchersByArea}
\title{GetResearchersByArea}
\usage{
GetResearchersByArea(ids.researcher, production.by.researcher)
}
\arguments{
\item{ids.researcher}{a list of Lattes CV XML files.}

\item{production.by.researcher}{field a list of Lattes CV XML files.}
}
\description{
GetResearchersByArea
}
