% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientacoes_desta_pessoa.R
\name{OrientacoesDestaPessoa}
\alias{OrientacoesDestaPessoa}
\title{OrientacoesDestaPessoa}
\usage{
OrientacoesDestaPessoa(id, orientacoesPorAno)
}
\arguments{
\item{id}{a list of Lattes CV XML files.}

\item{orientacoesPorAno}{field a list of Lattes CV XML files.}
}
\description{
OrientacoesDestaPessoa
}
