% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parserXML.R
\name{ParserXML}
\alias{ParserXML}
\title{parserXML}
\usage{
ParserXML(files = list())
}
\arguments{
\item{files}{a list of Lattes CV XML files}
}
\value{
a hash table indexed by id_LATTES of each researcher where each element corresponds to a data struct  AUTOR_DO_CURRICULO containing all the academic production of the researcher.
}
\description{
This function convert a set of Lattes CV XML files to a hash table indexed by #'id_LATTES each researcher
}
\examples{
ParserXML(files.selected)

}
