% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_one_json_perfil.R
\name{PartOneJsonPerfil}
\alias{PartOneJsonPerfil}
\title{PartOneJsonPerfil}
\usage{
PartOneJsonPerfil(publicacoes)
}
\arguments{
\item{publicacoes}{a list of Lattes CV XML files}
}
\description{
PartOneJsonPerfil
}
