% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicaoes_desta_pessoa.R
\name{PublicaoesDestaPessoa}
\alias{PublicaoesDestaPessoa}
\title{PublicaoesDestaPessoa}
\usage{
PublicaoesDestaPessoa(id, publicacoesPorAno)
}
\arguments{
\item{id}{a list of Lattes CV XML files.}

\item{publicacoesPorAno}{field a list of Lattes CV XML files.}
}
\description{
PublicaoesDestaPessoa
}
