% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{Save}
\alias{Save}
\title{Save}
\usage{
Save(content, path, filename)
}
\arguments{
\item{content}{a list of Lattes CV XML files.}

\item{path}{field a list of Lattes CV XML files.}

\item{filename}{field a list of Lattes CV XML files.}
}
\description{
Save
}
