% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_by_year.R
\name{SeparateByYear}
\alias{SeparateByYear}
\title{SeparateByYear}
\usage{
SeparateByYear(ids, h, parametros, tipo, tipoDaProducao)
}
\arguments{
\item{ids, }{h, parametros, tipo, tipoDaProducao  a list of Lattes CV XML files}
}
\description{
SeparateByYear
}
