% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_xml_attr.R
\name{ValidateXmlAttr}
\alias{ValidateXmlAttr}
\title{ValidateXmlAttr}
\usage{
ValidateXmlAttr(attr, field)
}
\arguments{
\item{attr, }{field a list of Lattes CV XML files}
}
\description{
ValidateXmlAttr
}
