% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_json.R
\name{BuildJson}
\alias{BuildJson}
\title{BuildJson}
\usage{
BuildJson(pessoa, publicacoes, orientacoes, senioridade)
}
\arguments{
\item{pessoa, }{publicacoes, orientacoes, senioridade}
}
\description{
BuildJson
}
