% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CREATE_GRAPH.R
\name{CreateGraph}
\alias{CreateGraph}
\title{CreateGraph}
\usage{
CreateGraph(profiles, publications)
}
\arguments{
\item{profiles}{json of profiles and publications.}

\item{publications}{field a list of Lattes CV XML files.}
}
\value{
return graphs
}
\description{
This function convert a set of Lattes CV XML files to a hash table indexed by #'id_LATTES each researcher
}
