% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{Execute}
\alias{Execute}
\title{Execute}
\usage{
Execute(ids, dado, conf, tipo, doTipo, levenshtein)
}
\arguments{
\item{ids}{a list of Lattes CV XML files.}

\item{dado}{field a list of Lattes CV XML files.}

\item{conf}{field a list of Lattes CV XML files.}

\item{tipo}{field a list of Lattes CV XML files.}

\item{doTipo}{field a list of Lattes CV XML files.}

\item{levenshtein}{field a list of Lattes CV XML files.}
}
\description{
Execute
}
