% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_production_by_year_and_type.R
\name{GetProductionByYearAndType}
\alias{GetProductionByYearAndType}
\title{GetProductionByYearAndType}
\usage{
GetProductionByYearAndType(conf.JSON, type.production, default, ids, data,
  levenshtein)
}
\arguments{
\item{conf.JSON, }{type.production, default, ids, data, levenshtein a list of Lattes CV XML files}
}
\description{
GetProductionByYearAndType
}
