% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type_production_no_duplicate.R
\name{GetTypeProductionNoDuplicate}
\alias{GetTypeProductionNoDuplicate}
\title{GetTypeProductionNoDuplicate}
\usage{
GetTypeProductionNoDuplicate(publicacoesPorAno, levenshtein, tipo,
  tipoDaProducao)
}
\arguments{
\item{publicacoesPorAno, }{levenshtein, tipo, tipoDaProducao a list of Lattes CV XML files}
}
\description{
GetTypeProductionNoDuplicate
}
