% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_two_json_perfil.R
\name{PartTwoJsonPerfil}
\alias{PartTwoJsonPerfil}
\title{PartTwoJsonPerfil}
\usage{
PartTwoJsonPerfil(publicacoes, tipo)
}
\arguments{
\item{publicacoes, tipo}{a list of Lattes CV XML files}
}
\description{
PartTwoJsonPerfil
}
