% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_by_year.R
\name{SeparateByYear}
\alias{SeparateByYear}
\title{SeparateByYear}
\usage{
SeparateByYear(ids, h, parametros, tipo, tipoDaProducao)
}
\arguments{
\item{ids}{a list of Lattes CV XML files.}

\item{h}{field a list of Lattes CV XML files.}

\item{parametros}{field a list of Lattes CV XML files.}

\item{tipo}{field a list of Lattes CV XML files.}

\item{tipoDaProducao}{field a list of Lattes CV XML files.}
}
\description{
SeparateByYear
}
